+++ iguana.taskprocessor/src/test/java/org/aksw/iguana/tp/tasks/impl/stresstest/StresstestTest.java
@@ -44,7 +44,7 @@ public class StresstestTest {
 	private static Connection connection;
 	private static Channel channel;
 	private String taskID = "test";
-	private Long timeLimit = null;
+	private Double timeLimit = null;
 	private Long noOfQueryMixes = 1l;
 	private String[] queryHandler = new String[] { "org.aksw.iguana.tp.query.impl.InstancesQueryHandler" };
 	private Object[][] workerConfigurations;
@@ -99,7 +99,7 @@ public class StresstestTest {
 	 * @param workerConfigurations
 	 * 
 	 */
-	public StresstestTest(String taskID, String service, String updateService, Long timeLimit, Long noOfQueryMixes, String[] queryHandler,
+	public StresstestTest(String taskID, String service, String updateService, Double timeLimit, Long noOfQueryMixes, String[] queryHandler,
 			Object[][] workerConfigurations) {
 		this.taskID = taskID;
 		this.timeLimit = timeLimit;
